2020.07.27 11:33 AM Code Operation Debug #2 TODO : 1) Init FIFO 2) VLCGroup()

+++ init_enc_params.py
@@ -182,7 +182,7 @@ class initDscConstants:
             self.numComponents = 3
 
         # range_ = np.zeros(self.NUM_COMPONENTS, )
-        self.cpntBitDepth = np.zeros(defines.NUM_COMPONENTS, )
+        self.cpntBitDepth = np.zeros(defines.NUM_COMPONENTS, ).astype(np.int32)
         for i in range(defines.NUM_COMPONENTS):
             self.cpntBitDepth[i] = 1 << pps.bits_per_component
             diff_cond = (pps.convert_rgb & (i is not 0) & (i is not 3) & (pps.bits_per_component is not 0))  # 16 bpc condition
